+++ src/Oro/Bundle/ApiBundle/ApiDoc/SecurityContext.php
@@ -2,6 +2,7 @@
 
 namespace Oro\Bundle\ApiBundle\ApiDoc;
 
+use Oro\Bundle\SecurityBundle\Authentication\Token\OrganizationContextTokenInterface;
 use Oro\Bundle\UserBundle\Security\AdvancedApiUserInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
@@ -66,6 +67,15 @@ class SecurityContext implements SecurityContextInterface
         $apiKeyKeys = $user->getApiKeys();
         if ($apiKeyKeys->isEmpty()) {
             return null;
+        }
+
+        if ($token instanceof OrganizationContextTokenInterface) {
+            $organization = $token->getOrganizationContext();
+            foreach ($apiKeyKeys as $apiKeyKey) {
+                if ($apiKeyKey->getOrganization()->getId() === $organization->getId()) {
+                    return $apiKeyKey->getApiKey();
+                }
+            }
         }
 
         return $apiKeyKeys->first()->getApiKey();